backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
django-storages
Pillow
boto3
#celery==5.2.7
#redis
#rabbitmq
djangorestframework-simplejwt
bcrypt
